+++ lib/src/model/platform/platform_overwrite_enum.dart
@@ -1,0 +1,8 @@
+enum PlatformOverwrite {
+  android,
+  iOS,
+  web,
+  macos,
+  linux,
+  windows,
+}